% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!
This function says hello to any person!}
\usage{
hello(name = "Felicity")
}
\arguments{
\item{name}{A string of characters, preferentially a name of a person. (The
default= Felicity)}
}
\value{
Prints Hello 'name' to the screen
}
\description{
Hello, world!
This function says hello to any person!
}
\examples{
Hello("Malie")
}
